abc133

+++ src/abc133/rain_flows_into_dams.h
@@ -1,1 +1,2 @@
-std::vector<int> rain_flows_into_dams(std::vector<int> a);
+
+std::vector<long long> rain_flows_into_dams(std::vector<long long> a);
